Refactor NmGraph extraction
